requirements.txt
# Core Dependencies
flask==2.0.1
flask-cors==3.0.10
pymongo==4.1.1
python-dotenv==0.19.0
requests==2.26.0

# Machine Learning Dependencies
numpy==1.21.2
pandas==1.3.3
scikit-learn==0.24.2
tensorflow==2.6.0
nltk==3.6.3
textblob==0.15.3

# Data Processing
pandas==1.3.3
numpy==1.21.2
matplotlib==3.4.3
seaborn==0.11.2

# API and Web
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
aiohttp==3.8.1

# Testing
pytest==6.2.5
pytest-cov==2.12.1
pytest-asyncio==0.15.1

# Development
black==21.7b0
flake8==3.9.2
isort==5.9.3
mypy==0.910 